# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition"
  },
...
  "paths" : {
...
    "/v1/rights" : {
      "post" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "file",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Request"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rights/{requestId}/fulfilment" : {
      "get" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "fulfilment",
        "parameters" : [ {
          "name" : "requestId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "recordFulfilment",
        "parameters" : [ {
          "name" : "requestId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FulfilmentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
    "/v1/rights/{requestId}/acknowledge" : {
      "post" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "acknowledge",
        "parameters" : [ {
          "name" : "requestId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Request"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/consent" : {
      "post" : {
        "tags" : [ "consent-controller" ],
        "operationId" : "capture",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CaptureRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CaptureResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/rights/{requestId}" : {
      "get" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "requestId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Request"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "transition",
        "parameters" : [ {
          "name" : "requestId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TransitionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Request"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/rights/summary" : {
      "get" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "summary",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TypeSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/rights/subject/{entityId}/{subjectId}" : {
      "get" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "forSubject",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subjectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Request"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/rights/queue" : {
      "get" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "queue",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 100
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Request"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/rights/overdue" : {
      "get" : {
        "tags" : [ "rights-controller" ],
        "operationId" : "overdue",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 200
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Request"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/receipts" : {
      "get" : {
        "tags" : [ "receipt-controller" ],
        "operationId" : "forSubject_1",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subjectId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StoredReceipt"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/provenance/{entityId}/{subjectId}" : {
      "get" : {
        "tags" : [ "provenance-controller" ],
        "operationId" : "forSubject_2",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "subjectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Record"
                }
              }
            }
          }
        }
      }
    },
    "/v1/provenance/summary" : {
      "get" : {
        "tags" : [ "provenance-controller" ],
        "operationId" : "summary_1",
        "parameters" : [ {
          "name" : "entityId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SourceSummary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "FileRequest" : {
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "subjectId" : {
            "type" : "string"
          },
          "identifierType" : {
            "type" : "string",
            "enum" : [ "PHONE", "EMAIL", "EMPLOYEE_ID", "CANDIDATE_ID", "DEVICE_ID", "EXTERNAL_ID" ]
          },
          "identifierValue" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "ACCESS", "CORRECTION", "COMPLETION", "ERASURE", "NOMINATION", "GRIEVANCE", "CONSENT_WITHDRAWAL", "PORTABILITY", "OPT_OUT_OF_SALE" ]
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "receivedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "details" : {
            "type" : "string"
          },
          "verifiedAs" : {
            "type" : "string"
          }
        },
        "required" : [ "entityId", "type" ]
      },
      "Request" : {
        "type" : "object",
        "properties" : {
          "requestId" : {
            "type" : "string"
          },
          "entityId" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "ACCESS", "CORRECTION", "COMPLETION", "ERASURE", "NOMINATION", "GRIEVANCE", "CONSENT_WITHDRAWAL", "PORTABILITY", "OPT_OUT_OF_SALE" ]
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "status" : {
            "type" : "string",
            "enum" : [ "RECEIVED", "IN_PROGRESS", "AWAITING_SUBJECT", "FULFILLED", "REJECTED", "WITHDRAWN" ]
          },
          "receivedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dueAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dueAtBasis" : {
            "type" : "string"
          },
          "closedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "acknowledgedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "assignedTo" : {
            "type" : "string"
          },
          "resolution" : {
            "type" : "string"
          },
          "details" : {
            "type" : "string"
          },
          "verification" : {
            "type" : "string",
            "enum" : [ "PORTAL_TOKEN", "OPERATOR_ASSERTED", "UNVERIFIED" ]
          },
          "verifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "verificationDetail" : {
            "type" : "string"
          }
        }
      },
      "FulfilmentRequest" : {
        "type" : "object",
        "properties" : {
          "systemCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "actionType" : {
            "type" : "string",
            "minLength" : 1
          },
          "status" : {
            "type" : "string",
            "minLength" : 1
          },
          "evidenceRef" : {
            "type" : "string",
            "minLength" : 1
          },
          "detail" : {
            "type" : "string"
          }
        },
        "required" : [ "actionType", "evidenceRef", "status", "systemCode" ]
      },
...
      "CaptureRequest" : {
        "type" : "object",
        "properties" : {
          "entityId" : {
            "type" : "string",
            "minLength" : 1
          },
          "subjectId" : {
            "type" : "string"
          },
          "subject" : {
            "$ref" : "#/components/schemas/SubjectRef"
          },
          "alsoKnownAs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SubjectRef"
            }
          },
          "jurisdiction" : {
            "type" : "string",
            "enum" : [ "IN", "EU", "UK", "SG", "MY", "KR", "US_CA", "US_CO", "US_CT", "US_TX", "US_OR", "US_MT", "US_DE", "US_NJ", "US_NE", "US_NH", "US_MN", "US_MD", "US_VA", "US_UT", "US_IA", "OTHER" ]
          },
          "languageTag" : {
            "type" : "string",
            "minLength" : 1
          },
          "channel" : {
            "type" : "string",
            "enum" : [ "WEB", "MOBILE_APP", "VOICE_CALL", "SMS", "WHATSAPP", "EMAIL", "KIOSK", "PAPER", "API", "BULK_IMPORT", "POSTAL" ]
          },
          "applicationId" : {
            "type" : "string"
          },
          "captureMethod" : {
            "type" : "string",
            "enum" : [ "CHECKBOX_OPT_IN", "CLICK_THROUGH", "DOUBLE_OPT_IN", "OTP_VERIFIED", "VERBAL_RECORDED", "WET_SIGNATURE", "BIOMETRIC_ENROLMENT", "PARENTAL_VERIFIED", "RELAYED_BY_CONSENT_MANAGER", "IMPORTED_WITH_PROVENANCE", "INFERRED_FROM_RELATIONSHIP", "NOT_APPLICABLE" ]
          },
          "actorType" : {
            "type" : "string",
            "enum" : [ "SUBJECT", "PARENT_GUARDIAN", "ADMIN", "AGENT", "SYSTEM", "IMPORT", "REGULATOR", "CONSENT_MANAGER" ]
          },
          "actorId" : {
            "type" : "string"
          },
          "noticeId" : {
            "type" : "string"
          },
          "noticeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "choices" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PurposeChoiceDto"
            },
            "minItems" : 1
          },
          "rejectAllOffered" : {
            "type" : "boolean"
          },
          "occurredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "idempotencyKey" : {
            "type" : "string"
          },
          "evidenceRef" : {
            "type" : "string"
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "guardianVerification" : {
            "$ref" : "#/components/schemas/GuardianVerificationDto"
          }
        },
        "required" : [ "actorType", "captureMethod", "choices", "entityId", "jurisdiction", "languageTag" ]
      },
...
      "PurposeChoiceDto" : {
        "type" : "object",
        "properties" : {
          "purposeCode" : {
            "type" : "string",
            "minLength" : 1
          },
          "granted" : {
            "type" : "boolean"
          },
          "preTicked" : {
            "type" : "boolean"
          },
          "separateAction" : {
            "type" : "boolean"
          }
        },
        "required" : [ "purposeCode" ]
      },
      "SubjectRef" : {
        "type" : "object",
        "properties" : {
          "identifierType" : {
            "type" : "string",
            "enum" : [ "PHONE", "EMAIL", "EMPLOYEE_ID", "CANDIDATE_ID", "DEVICE_ID", "EXTERNAL_ID" ]
          },
          "value" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "identifierType", "value" ]
      },
      "CaptureResponse" : {
        "type" : "object",
        "properties" : {
          "accepted" : {
            "type" : "boolean"
          },
          "subjectId" : {
            "type" : "string"
          },
          "events" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EventSummary"
            }
          },
          "violations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ViolationDto"
            }
          }
        }
      },
...
      "Record" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "entityId" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string"
          },
          "sourceType" : {
            "type" : "string"
          },
          "sourceName" : {
            "type" : "string"
          },
          "acquiredAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "originalLegalBasis" : {
            "type" : "string"
          },
          "contractRef" : {
            "type" : "string"
          },
          "substantiated" : {
            "type" : "boolean"
          },
          "quarantined" : {
            "type" : "boolean"
          }
        }
      },
...
      "TransitionRequest" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "RECEIVED", "IN_PROGRESS", "AWAITING_SUBJECT", "FULFILLED", "REJECTED", "WITHDRAWN" ]
          },
          "assignedTo" : {
            "type" : "string"
          },
          "resolution" : {
            "type" : "string"
          }
        },
        "required" : [ "status" ]
      },
...
      "TypeSummary" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "enum" : [ "ACCESS", "CORRECTION", "COMPLETION", "ERASURE", "NOMINATION", "GRIEVANCE", "CONSENT_WITHDRAWAL", "PORTABILITY", "OPT_OUT_OF_SALE" ]
          },
          "open" : {
            "type" : "integer",
            "format" : "int64"
          },
          "overdue" : {
            "type" : "integer",
            "format" : "int64"
          },
          "closed" : {
            "type" : "integer",
            "format" : "int64"
          },
          "closedLate" : {
            "type" : "integer",
            "format" : "int64"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "StoredReceipt" : {
        "type" : "object",
        "properties" : {
          "receiptId" : {
            "type" : "string"
          },
          "entityId" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string"
          },
          "issuedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "payload" : {
            "type" : "string"
          },
          "payloadHash" : {
            "type" : "string"
          },
          "evidenceHash" : {
            "type" : "string"
          },
          "noticeId" : {
            "type" : "string"
          },
          "noticeVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "languageTag" : {
            "type" : "string"
          },
          "purposeCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "SourceSummary" : {
        "type" : "object",
        "properties" : {
          "sourceType" : {
            "type" : "string"
          },
          "sourceName" : {
            "type" : "string"
          },
          "quarantined" : {
            "type" : "integer",
            "format" : "int64"
          },
          "contactable" : {
            "type" : "integer",
            "format" : "int64"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    }
  }
}